Rename classes and functions
